backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
openai
langchain
langchain-core
langchain-openai
langchain-groq
langchain-google-genai
langchain-anthropic
databricks-langchain
langchain-ollama
tavily-python
gtts
elevenlabs
icalendar
httpx
python-multipart
pypdf
python-docx
python-pptx
requests
pydantic